refactor: split floating-ui callout renderer
Keep the default callout behavior as a simple red fallback bubble and move the collision-aware Floating UI layout into an explicit optional renderer. Refresh the Playwright demo GIF and update runtime watcher coverage for the new helper.
