+++ plugin/support/build-plugin.ts
@@ -348,6 +348,11 @@ export function createBuildProcessorPlugin(options: BuildProcessorOptions): Plug
         this.error(`callout.ts not found at ${calloutPath}. Ensure it is included in the build.`);
       }
       this.addWatchFile(calloutPath);
+
+      const floatingUiCalloutPath = path.resolve(path.dirname(basePageClassPath), "floating-ui-callout.ts");
+      if (fs.existsSync(floatingUiCalloutPath)) {
+        this.addWatchFile(floatingUiCalloutPath);
+      }
     },
     async buildEnd(error) {
       if (error) {
